+++ formset/richtext/controls.py
@@ -1,7 +1,7 @@
 import re
 
 from django.core.exceptions import ImproperlyConfigured
-from django.template.loader import select_template
+from django.template.loader import get_template, select_template
 from django.utils.functional import cached_property
 from django.utils.translation import gettext_lazy as _
 
@@ -36,6 +36,18 @@ class ControlElement:
         template = self.get_template(renderer)
         if context is None:
             context = self.get_context()
+        return template.render(context)
+
+
+class Group(list):
+    template_name = 'formset/richtext/control_group.html'
+
+    def render(self, renderer, context=None):
+        if context is None:
+            context = {
+                'elements': [element.render(renderer) for element in self],
+            }
+        template = get_template(self.template_name)
         return template.render(context)
 
 
@@ -124,6 +136,28 @@ class TextColor(ControlElement):
         return template.render({
             'colors': self.colors,
             'class_based': self.class_based,
+        })
+
+
+class FontFamily(ControlElement):
+    extension = 'fontFamily'
+    label = _("Font Family")
+    template_name = 'formset/richtext/font.html'
+
+    def __init__(self, css_classes):
+        if not isinstance(css_classes, dict) or len(css_classes) == 0:
+            raise ImproperlyConfigured("FontFamily() requires a dict with at least one entry")
+        class_pattern = re.compile(r'^-?[_a-zA-Z]+[_a-zA-Z0-9-]*$')
+        for css_class in css_classes.keys():
+            if not re.match(class_pattern, css_class):
+                raise ImproperlyConfigured(f"Given color {css_class} does not look like a valid CSS class name")
+        self.css_classes = {None: _("Default")}  # the default font
+        self.css_classes.update(css_classes)
+
+    def render(self, renderer):
+        template = self.get_template(renderer)
+        return template.render({
+            'css_classes': self.css_classes,
         })